[#1] Renamed adapter to contextLoader
